# backend/docs/swagger.yaml
basePath: /
definitions:
...
  services.RouteItem:
    properties:
      children:
        items:
          $ref: '#/definitions/services.RouteItem'
        type: array
      component:
        type: string
      meta:
        $ref: '#/definitions/services.RouteMeta'
      name:
        type: string
      path:
        type: string
    type: object
  services.RouteMeta:
    properties:
      activeIcon:
        type: string
      authority:
        items:
          type: integer
        type: array
      darkIcon:
        type: string
      icon:
        type: string
      order:
        type: integer
      title:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: 租户中心系统的API文档
  title: 租户中心API
  version: "1.0"
paths:
...
  /api/login:
    post:
      consumes:
      - application/json
      description: 用户登录接口，验证用户名和密码，返回JWT token
      parameters:
      - description: 登录信息
        in: body
        name: loginData
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功，返回token
          schema:
            type: object
        "400":
          description: 无效的请求参数
          schema:
            type: object
        "401":
          description: 用户名或密码错误
          schema:
            type: object
      tags:
      - 用户管理
...
  /api/users:
    post:
      consumes:
      - application/json
      description: 创建新用户，需要管理员权限
      parameters:
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: 用户创建成功
          schema:
            type: object
        "400":
          description: 无效的请求参数
          schema:
            type: object
        "500":
          description: 创建用户失败
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      tags:
      - 用户管理
  /api/users/{id}:
    put:
      consumes:
      - application/json
      description: 更新指定用户的信息，需要管理员权限
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 用户信息更新成功
          schema:
            type: object
        "400":
          description: 无效的请求参数
          schema:
            type: object
        "500":
          description: 更新用户信息失败
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      tags:
      - 用户管理
  /api/users/{id}/roles:
    post:
      consumes:
      - application/json
      description: 为指定用户绑定一个或多个角色，需要管理员权限
      parameters:
      - description: 用户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 角色ID列表
        in: body
        name: roleIDs
        required: true
        schema:
          items:
            type: integer
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: 角色绑定成功
          schema:
            type: object
        "400":
          description: 无效的请求参数
          schema:
            type: object
        "500":
          description: 绑定角色失败
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      tags:
      - 用户管理
  /api/users/routes:
    get:
      description: 获取用户的路由数据，包括菜单和权限信息
      produces:
      - application/json
      responses:
        "200":
          description: 路由数据
          schema:
            items:
              $ref: '#/definitions/services.RouteItem'
            type: array
        "400":
          description: 无效的用户ID
          schema:
            type: object
        "500":
          description: 获取路由数据失败
          schema:
            type: object
      security:
      - ApiKeyAuth: []
      tags:
      - 用户管理
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"